StyleWard/requirements.txt
absl-py==2.0.0
altair==5.2.0
astor==0.8.1
asttokens==2.4.1
attrs==23.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
Flask==3.0.0
fonttools==4.46.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.40
graphviz==0.20.1
grpcio==1.59.3
h5py==3.10.0
idna==3.6
imbalanced-learn==0.11.0
importlib-metadata==6.11.0
ipykernel==6.27.1
ipython==8.18.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter_client==8.6.0
jupyter_core==5.5.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
nest-asyncio==1.5.8
numpy==1.26.2
packaging==23.2
pandas==2.1.3
parso==0.8.3
Pillow==10.1.0
platformdirs==4.0.0
playsound==1.2.2
prompt-toolkit==3.0.41
protobuf==4.25.1
psutil==5.9.6
pure-eval==0.2.2
pyarrow==14.0.1
pydeck==0.8.1b0
pygame==2.5.2
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
pyzmq==25.1.1
referencing==0.31.1
requests==2.31.0
rich==13.7.0
rpds-py==0.13.2
scikit-learn==1.3.2
scipy==1.11.4
setuptools==69.0.2
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
streamlit==1.29.0
tenacity==8.2.3
tensorboard==1.12.2
tensorflow @ https://storage.googleapis.com/tensorflow/mac/cpu/tensorflow-1.12.0-py3-none-any.whl#sha256=f587dc03b5f0d1e50cca39b7159c9f21ffdec96273dbf5f7619d48c622cb21f2
termcolor==2.4.0
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.4
traitlets==5.14.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.12
Werkzeug==3.0.1
wheel==0.42.0
zipp==3.17.0
